+++ lib/ui/component/modal_sheet/app_detail_master_modal_sheet.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:food_gram_app/core/model/posts.dart';
 import 'package:food_gram_app/core/model/users.dart';
-import 'package:food_gram_app/core/utils/common_dialog.dart';
+import 'package:food_gram_app/core/utils/helpers/dialog_helper.dart';
 import 'package:food_gram_app/gen/l10n/l10n.dart';
 import 'package:food_gram_app/ui/screen/detail/detail_post_view_model.dart';
 import 'package:go_router/go_router.dart';
@@ -58,7 +58,7 @@ class AppDetailMasterModalSheet extends ConsumerWidget {
                     ),
                   ),
                   onPressed: () {
-                    openDialog(
+                    DialogHelper().openDialog(
                       title: 'マスター権限の投稿削除',
                       text: '不正な投稿と自分の投稿を削除するために使用します',
                       onTap: () async {
